Refactor tests to reduce code duplication.
There were excessive definitions of modules and datasets. More code is
moved to conftest now.
